repo reorg: move mongo filter
