+++ source/extensions/all_extensions.bzl
@@ -6,6 +6,7 @@ def envoy_all_extensions(repository = ""):
   return [
     repository + "//source/extensions/filters/network/client_ssl_auth:config",
     repository + "//source/extensions/filters/network/echo:config",
+    repository + "//source/extensions/filters/network/mongo_proxy:config",
     repository + "//source/extensions/filters/network/tcp_proxy:config",
   ]
