+++ test/extensions/filters/network/mongo_proxy/BUILD
@@ -13,7 +13,7 @@ envoy_cc_test(
     srcs = ["bson_impl_test.cc"],
     deps = [
         "//source/common/buffer:buffer_lib",
-        "//source/common/mongo:bson_lib",
+        "//source/extensions/filters/network/mongo_proxy:bson_lib",
     ],
 )
 
@@ -23,8 +23,8 @@ envoy_cc_test(
     deps = [
         "//source/common/buffer:buffer_lib",
         "//source/common/json:json_loader_lib",
-        "//source/common/mongo:bson_lib",
-        "//source/common/mongo:codec_lib",
+        "//source/extensions/filters/network/mongo_proxy:bson_lib",
+        "//source/extensions/filters/network/mongo_proxy:codec_lib",
     ],
 )
 
@@ -33,15 +33,16 @@ envoy_cc_test(
     srcs = ["proxy_test.cc"],
     deps = [
         "//source/common/event:dispatcher_lib",
-        "//source/common/mongo:bson_lib",
-        "//source/common/mongo:codec_lib",
-        "//source/common/mongo:proxy_lib",
         "//source/common/stats:stats_lib",
+        "//source/extensions/filters/network/mongo_proxy:bson_lib",
+        "//source/extensions/filters/network/mongo_proxy:codec_lib",
+        "//source/extensions/filters/network/mongo_proxy:config",
         "//test/mocks/access_log:access_log_mocks",
         "//test/mocks/event:event_mocks",
         "//test/mocks/filesystem:filesystem_mocks",
         "//test/mocks/network:network_mocks",
         "//test/mocks/runtime:runtime_mocks",
+        "//test/mocks/server:server_mocks",
     ],
 )
 
@@ -49,8 +50,8 @@ envoy_cc_test(
     name = "utility_test",
     srcs = ["utility_test.cc"],
     deps = [
-        "//source/common/mongo:bson_lib",
-        "//source/common/mongo:codec_lib",
-        "//source/common/mongo:utility_lib",
+        "//source/extensions/filters/network/mongo_proxy:bson_lib",
+        "//source/extensions/filters/network/mongo_proxy:codec_lib",
+        "//source/extensions/filters/network/mongo_proxy:utility_lib",
     ],
 )
