+++ test/integration/BUILD
@@ -301,7 +301,6 @@ envoy_cc_test_library(
         "//source/server/config/http:router_lib",
         "//source/server/config/http:zipkin_lib",
         "//source/server/config/network:http_connection_manager_lib",
-        "//source/server/config/network:mongo_proxy_lib",
         "//source/server/config/network:ratelimit_lib",
         "//source/server/config/network:raw_buffer_socket_lib",
         "//source/server/config/network:redis_proxy_lib",
